NO-TICKET - use translationY instead margin
